move files into components/HexGrid/
