--- src/components/BaseInput.vue
@@ -1,3 +1,0 @@
-<template>
-  <input />
-</template>
